+++ src/api/projects/repository/files.rs
@@ -9,6 +9,7 @@
 //! These endpoints are used for querying a project's files.
 
 mod create;
+mod delete;
 mod file_raw;
 mod update;
 
@@ -16,6 +17,10 @@ pub use self::create::CreateFile;
 pub use self::create::CreateFileBuilder;
 pub use self::create::CreateFileBuilderError;
 pub use self::create::Encoding;
+
+pub use self::delete::DeleteFile;
+pub use self::delete::DeleteFileBuilder;
+pub use self::delete::DeleteFileBuilderError;
 
 pub use self::file_raw::FileRaw;
 pub use self::file_raw::FileRawBuilder;